data: update the SVGs to put a proper anchor for the texts
If the anchor of the text is on the left but the leader on the right,
there will be artefacts when changing the name. Make sure all of the
labels are prooperly set.

This forced us to also relocate the labels on the G500s and the etekcity
scroll alpha.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>