test: fix the loading of the test driver
This test driver is already twisting a little bit the API, so twist it
a little bit more.

An other solution would be to just add a (void* data) parameter to .probe()
but there is no point in it because only test_driver would be used.

Create a new probe callback only used when using the test driver, so make
check is happy.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>