Add async commit support
Defer the actual commit so we can return immediately. This fixes the various
timeout issues we've seen on some devices where a commit can take several
seconds.

This commit merely adds the async hook. If the commit itself fails we
should re-sync from libratbag and update all the properties. This is currently
missing.

Fixes #269

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>