Drop the modifiers argument from the key action type
Any device that could do modifiers should be able to do macros, and where not
we can probably emulate macros *within that device's driver*. Meanwhile,
assigning a single key to a button is a common feature and we should expose
this for simplicity. UIs like piper could wrap this into the macro recording
UI where needed.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>